Update Comparing Number.cs
Just added some comments for the horrible code of using what if else. This could've been better if for loops with arrays were utilized in this scenario. 

For ex. for (int i = 0; i < sets.Length; i++) then determine the biggest, smallest with that.

Or another example using method/function

Make the code cleaner with that.

For ex. 

static void FindBigSmallNumber(int a, int b, int c)

Then, put the rest of the what if horrid code there.
Using the initialized variable inside the method.

"Do not overcomplicate everything, Be simplistic" - Emmanuel